sva-73: split feature into multiple files
