+++ src/pushNotifications/index.ts
@@ -1,0 +1,2 @@
+export * from './PermissionHandling'
+export * from './TokenHandling'
